fix: fixing mismatched data type
* In schema.ts, salesInvoiceItems.quantity is a numeric type, so we modify qtySold that select quantity to `::numeric`.
* Another way we could tackle this problem is by removing the cast altogether

+++ frontend/src/server/reports.ts
@@ -227,7 +227,7 @@ export async function getSalesReportForPeriod(period: SalesReportPeriod) {
       : await db
           .select({
             productName: salesInvoiceItems.productName,
-            qtySold: sql<number>`sum(${salesInvoiceItems.quantity})::integer`,
+            qtySold: sql<number>`sum(${salesInvoiceItems.quantity})::numeric`,
             revenue: sql<number>`sum(${salesInvoiceItems.lineTotal})::integer`,
           })
           .from(salesInvoiceItems)
